+++ src/main.cpp
@@ -5,7 +5,7 @@ int main(int argc, char *argv[])
 {
     QApplication app(argc, argv);
     app.setApplicationName("BlockTorch");
-    app.setApplicationVersion("0.0.3");
+    app.setApplicationVersion("0.0.4");
 
     MainWindow window;
     window.show();