+++ src/ui/HomeView.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <QWidget>
+
+class HomeView : public QWidget {
+    Q_OBJECT
+
+public:
+    explicit HomeView(QWidget *parent = nullptr);
+
+signals:
+    void workspaceRequested();
+};