+++ src/ui/MainWindow.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <QMainWindow>
+#include <QStackedWidget>
 
 class MainWindow : public QMainWindow {
     Q_OBJECT
@@ -12,4 +13,9 @@ public:
 private:
     void setupTheme();
     void setupUI();
+
+    QStackedWidget *m_stacked;
+    int m_homeIndex;
+    int m_savesIndex;
+    int m_workspaceIndex;
 };